Migrate from patchset to contrib
